+++ config.json
@@ -1,8 +1,8 @@
 {
-  "port_service": "127.0.0.1:8382",
+  "Url_service": "127.0.0.1:8382",
   "sms_path": "C:/Users/User/go/src/skillbox-diploma/sms.data",
   "mms_url":"http://127.0.0.1:8383/mms",
-  "vice_call_path":"C:/Users/User/go/src/skillbox-diploma/voice.data",
+  "voice_call_path":"C:/Users/User/go/src/skillbox-diploma/voice.data",
   "email_path":"C:/Users/User/go/src/skillbox-diploma/email.data",
   "billing_path":"C:/Users/User/go/src/skillbox-diploma/billing.data",
   "incident_url":"http://127.0.0.1:8383/accendent",
